removed export method

+++ src/JavaNs2/Models/Helper.java
@@ -1,12 +1,6 @@
 package JavaNs2.Models;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.Calendar;
-import java.util.Date;
 
 public class Helper {
     private static long startTime;
@@ -28,7 +22,7 @@ public class Helper {
     public static void endPerformanceTest()
     {
         long endTime = System.currentTimeMillis();
-        long execTime = (endTime - startTime) / 1000;
+        float execTime = (float)(endTime - startTime) / 1000;
         System.out.println("End at: " + endTime);
         System.out.println("Performance: " + execTime + "s");
     }
